[DEBUG] fix frame timestamps: 64 bits value
Reviewed-on: https://gerrit.parrot.biz/49680

+++ JNI/java/com/parrot/arsdk/arcontroller/ARFrame.java
@@ -10,7 +10,7 @@ public class ARFrame extends ARNativeData
     private int missed;
     private boolean isIFrame;
     private boolean available;
-    private int timestamp;
+    private long timestamp;
     private ARNativeData metadata;
     
     public ARFrame ()
@@ -26,7 +26,7 @@ public class ARFrame extends ARNativeData
     }
     
     public ARFrame (long data, int dataCapacity, int dataSize, boolean isIFrame, int missed,
-                    int timestamp, long metadata, int metadataSize)
+                    long timestamp, long metadata, int metadataSize)
     {
         super(data, dataCapacity);
         initialize ();
@@ -85,7 +85,7 @@ public class ARFrame extends ARNativeData
         this.available = available;
     }
 
-    public int getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
